Adding my vimrc to the repo
He was not impressed with the response from the others, as they had seen the others in the cave and they were too young to have been the first ones there. "I see it now," he said, this is an unicorn!  You are a new breed of Unicorn , an untamed Wildebeest, a unicorn of the earth. You are the ones in the sky, the ones who walk among the stars. The others were still not convinced and thought that it was a myth and so said no. He took the new breed in the direction of the stars, to the north. There, the earth seemed empty and empty was all that there was.